--- src/utils/asyncTimeout.ts
@@ -1,3 +1,0 @@
-export function timeout(ms: number): Promise<void> {
-  return new Promise(resolve => setTimeout(resolve, ms))
-}
