--- src/utils/getRandomStringFromArray.ts
@@ -1,5 +1,0 @@
-export function getRandomStringFromArray(arr: string[]): string {
-  const length = arr.length
-  const randomNum = Math.floor(Math.random() * length)
-  return arr[randomNum]
-}
